♻️ refactor Contributors table view
